code refactor/cleanup
