+++ wallet-sdk-core/src/androidMain/kotlin/com/michaeltchuang/walletsdk/core/railmpp/internal/AlgorandOps.android.kt
@@ -11,7 +11,7 @@ import com.algorand.algosdk.v2.client.common.AlgodClient
 import com.algorand.algosdk.v2.client.common.Response
 import com.algorand.algosdk.v2.client.model.PostTransactionsResponse
 import com.michaeltchuang.walletsdk.core.railmpp.AndroidMppWalletSigner
-import com.michaeltchuang.walletsdk.core.railmpp.MppWalletSigner
+import com.michaeltchuang.walletsdk.core.railmpp.domain.repository.MppWalletSigner
 import com.michaeltchuang.walletsdk.core.utils.GoMobileDispatcher
 import java.math.BigInteger
 import java.net.URI
