+++ wallet-sdk-core/src/commonMain/kotlin/com/michaeltchuang/walletsdk/core/railmpp/smartcontract/EscrowSessionVaultManagerClient.kt
@@ -1,7 +1,7 @@
 package com.michaeltchuang.walletsdk.core.railmpp.smartcontract
 
 import com.michaeltchuang.walletsdk.core.deeplink.utils.AssetConstants
-import com.michaeltchuang.walletsdk.core.railmpp.MppWalletSigner
+import com.michaeltchuang.walletsdk.core.railmpp.domain.repository.MppWalletSigner
 import com.michaeltchuang.walletsdk.core.railmpp.internal.decodeAlgorandAddressPublicKey
 import com.michaeltchuang.walletsdk.core.railmpp.internal.encodeAlgorandAddress
 import com.michaeltchuang.walletsdk.core.railmpp.internal.encodeUint64
